+++ src/Security/LogoutListener.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Bolt\Security;
 
+use Bolt\Entity\User;
 use Doctrine\Common\Persistence\ObjectManager;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -22,6 +23,9 @@ class LogoutListener implements LogoutHandlerInterface
     public function logout(Request $request, Response $response, TokenInterface $token): void
     {
         $user = $token->getUser();
+        if (! $user instanceof User) {
+            return;
+        }
         $this->em->remove($user->getUserAuthToken());
         $this->em->flush();
     }
